Translation_Management_System-main/requirements.txt
# Core Framework
fastapi==0.110.2
uvicorn[standard]==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# HTTP & API
httpx==0.27.0
requests==2.31.0
aiofiles==23.2.1

# LLM Integration
openai==1.6.1
anthropic==0.8.1
google-generativeai==0.3.2

# Background Tasks
celery==5.3.4
redis==5.0.1

# WebSocket
websockets==12.0

# File Processing
python-magic==0.4.27
python-docx==1.1.0
openpyxl==3.1.2
PyPDF2==3.0.1

# Email
fastapi-mail==1.4.1

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.1
isort==5.13.2
flake8==6.1.0

# Production
gunicorn==21.2.0
